Move defer from devtools to util
