+++ public/js/clients/firefox.js
@@ -2,7 +2,7 @@ const { DebuggerClient } = require("devtools-sham/shared/client/main");
 const { DebuggerTransport } = require("devtools-sham/transport/transport");
 const WebSocketDebuggerTransport = require("devtools-sham/transport/ws-transport");
 const { TargetFactory } = require("devtools-sham/client/framework/target");
-const defer = require("../lib/devtools/shared/defer");
+const defer = require("../utils/defer");
 const { getValue } = require("../feature");
 const { Tab } = require("../types");
 const { setupCommands, clientCommands } = require("./firefox/commands");
