--- public/js/lib/devtools/shared/defer.js
@@ -1,25 +1,0 @@
-/* This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
-
-"use strict";
-
-// See bug 1273941 to understand this choice of promise.
-const Promise = require("promise");
-
-/**
- * Returns a deferred object, with a resolve and reject property.
- * https://developer.mozilla.org/en-US/docs/Mozilla/JavaScript_code_modules/Promise.jsm/Deferred
- */
-module.exports = function defer() {
-  let resolve, reject;
-  let promise = new Promise(function () {
-    resolve = arguments[0];
-    reject = arguments[1];
-  });
-  return {
-    resolve: resolve,
-    reject: reject,
-    promise: promise
-  };
-};
